add MatchDto, PlayerDto, PlayerInPositionDto & ReplacementDto

+++ BackEnd-MatchesService/src/main/java/tn/iit/entity/Replacement.java
@@ -4,13 +4,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
-import java.math.BigInteger;
-
 @Data
 @Builder
 @Jacksonized
 public class Replacement {
-    private BigInteger replacementTime;
+    private int replacementTime;
     private Player playerIn;
     private Player playerOut;
 }
